+++ Changelog.md
@@ -4,6 +4,8 @@
 - **3.2.0-SNAPSHOT**
    - Modified quoting for command line arguments
    - Implemented subtitle ass video filter
+   - Added constructor for scaling filter which allows string expressions
+   - Added constructor for scaling filter which does not require the ForceOriginalAspectRatio parameter
 - **3.1.0**
    - Added support for arm32 bit (Thanks to jmformenti)
    - Added option to use a specific quote character for command line