+++ Changelog.md
@@ -6,6 +6,7 @@
    - Implemented subtitle ass video filter
    - Added constructor for scaling filter which allows string expressions
    - Added constructor for scaling filter which does not require the ForceOriginalAspectRatio parameter
+   - Added CropFilter
 - **3.1.0**
    - Added support for arm32 bit (Thanks to jmformenti)
    - Added option to use a specific quote character for command line