+++ Changelog.md
@@ -8,6 +8,8 @@
    - Added constructor for scaling filter which does not require the ForceOriginalAspectRatio parameter
    - Added CropFilter
    - Added constructor for color filter which allows string expressions
+   - Added support for multiple video filters in one conversion pass
+   - Added enhanced meta data detection in MultiMedia object
 - **3.1.0**
    - Added support for arm32 bit (Thanks to jmformenti)
    - Added option to use a specific quote character for command line